+++ container-engine-lib/lib/backend_impls/metrics_reporting/metrics_reporting_kurtosis_backend.go
@@ -199,30 +199,38 @@ func (backend *MetricsReportingKurtosisBackend) DestroyAPIContainers(ctx context
 
 func (backend *MetricsReportingKurtosisBackend) CreateModule(
 	ctx context.Context,
+	image string,
+	enclaveId enclave.EnclaveID,
 	id module.ModuleID,
 	guid module.ModuleGUID,
-	containerImageName string,
-	serializedParams string,
+	ipAddr net.IP, // TODO REMOVE THIS ONCE WE FIX THE STATIC IP PROBLEM!!
+	grpcPortNum uint16,
+	enclaveDataDirpathOnHostMachine string,
+	envVars map[string]string,
 )(
 	newModule *module.Module,
 	resultErr error,
 ) {
 	newModule, err := backend.underlying.CreateModule(
 		ctx,
+		image,
+		enclaveId,
 		id,
 		guid,
-		containerImageName,
-		serializedParams,
-		)
+		ipAddr,
+		grpcPortNum,
+		enclaveDataDirpathOnHostMachine,
+		envVars,
+	)
 	if err != nil {
 		return nil,
 		stacktrace.Propagate(
 			err,
-			"An error occurred creating module with ID '%v', GUID '%v', container image name '%v' and serialized params '%+v'",
+			"An error occurred creating module with ID '%v', GUID '%v', and image '%v'",
 			id,
 			guid,
-			containerImageName,
-			serializedParams)
+			image,
+		)
 	}
 
 	return newModule, nil
@@ -232,7 +240,7 @@ func (backend *MetricsReportingKurtosisBackend) GetModules(
 	ctx context.Context,
 	filters *module.ModuleFilters,
 )(
-	map[string]*module.Module,
+	map[module.ModuleGUID]*module.Module,
 	error,
 ) {
 	modules, err := backend.underlying.GetModules(ctx, filters)